feat(web): move jobs related questions in project form to their own chapter
